refactor: import 컴포넌트 케밥 케이스 적용 (#447)
